renames and moves some syllabus files and also adds the SyllabusContoller for file upload and download.  May need to do user-type checking on upload to make sure the user is the instructor and student's cannot upload files.
