+++ src/main/java/com/ako/api/SyllabusController.java
@@ -1,18 +1,44 @@
 package com.ako.api;
 
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.MediaType;
+import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.ResponseBody;
+
+import com.ako.service.SyllabusService;
 
 /**
- * @author Prashant
+ * Syllabus File Upload and Download Controller
+ * @author Tim
  */
 
-@RestController
-@RequestMapping( value = "/syllabus", produces = MediaType.APPLICATION_JSON_VALUE )
+@Controller
+@RequestMapping(value="/syllabus", produces=MediaType.APPLICATION_PDF_VALUE)
 public class SyllabusController {
+
+    @Autowired
+    SyllabusService syllabusService;
 	
-	public String testMethod() {
-		return "Hi";
+
+    /* File DOWNLOAD */
+    @RequestMapping(method=RequestMethod.GET, value="/")
+	public @ResponseBody void downloadSyllabus() {
+        
+		syllabusService.downloadFile("syllabus.pdf");
+
 	}
+
+
+    /* File UPLOAD    */
+    @RequestMapping(method=RequestMethod.POST, value="/")
+    public @ResponseBody void uploadSyllabus(@RequestParam("File") 
+                                                String file) {
+
+        syllabusService.uploadFile("syllabus.pdf", file);
+
+    }
+
 }
